requirements.txt
requests
google-generativeai
python-dotenv
openai
streamlit
feedparser
